The stripe has no inline shadow and it's smaller
The reasoning behind removing the inline shadow is that it's uneeded.
It's a distraction, and the overall theme wants to go the flat route.

The reasoning behind making the stripe smaller, is that when it's too
tall it's distracting, also 0.5 em seems to be a bit too much, especially
when compared against the logo thicker part ( 8px ). I've lowered it to half
of that 0.22em ~= 4px.