Rewrite the contribution section.
The high-level goal of this rewrite is to make it easier to contribute.
It focuses less on requirements, especially for tweaks that can easily
be added by Chris and I after the fact, and more on familiarizing
potential contributors with what to do and how to get help with it.