Avoid an unstyled rendering moment with code examples
This renders the tabs and all the jQuery classes server-side, so that
the code examples are rendered correctly on the page's first load
rather than waiting for the JavaScript to activate. This also has a
side effect of making them look correct with JavaScript disabled.

This also adds a little additional styling in the noscript stylesheet
to make code examples look better on a narrow viewport with JavaScript
disabled.