Update the banner to read "Free Palestine"
This change is an act of political speech proposed by me and agreed to
by the Sass core team. It does not reflect the opinions of my employer
or theirs.

As a Jew, I say: never again means never again. As an opponent of
imperialism and settler colonialism everywhere, I say: from the river
to the sea, Palestine will be free.